refactor(writer): route document IO through providers
