+++ lazyllm/tools/writer/provider/__init__.py
@@ -1,0 +1,19 @@
+from .base import WriterProviderBase
+from .feishu import FeishuWriterProvider
+from .registry import (
+    get_writer_provider,
+    match_writer_provider,
+    register_writer_provider,
+)
+
+
+register_writer_provider(FeishuWriterProvider)
+
+
+__all__ = [
+    'FeishuWriterProvider',
+    'WriterProviderBase',
+    'get_writer_provider',
+    'match_writer_provider',
+    'register_writer_provider',
+]
